chore: [form] organize test-related files into test folder
