+++ packages/components/form/__tests__/form-item.test.tsx
@@ -14,7 +14,7 @@ import { rAF } from '@element-plus/test-utils/tick'
 import Input from '@element-plus/components/input'
 import Form from '../src/form.vue'
 import FormItem from '../src/form-item.vue'
-import DynamicFormItem from '../mocks/mock-data'
+import DynamicFormItem from './mock-data'
 
 import type { VueWrapper } from '@vue/test-utils'
 import type { MockInstance } from 'vitest'
